chore: move local data source files on new directory
